Split up unit test files
