+++ test/reselect.spec.ts
@@ -123,7 +123,6 @@ describe('Basic selector behavior', () => {
     expect(selector.recomputations()).toBe(1)
     // Expected a million calls to a selector with the same arguments to take less than 1 second
     expect(totalTime).toBeLessThan(1000)
-    console.log('Total basic time: ', totalTime)
   })
 
   test('basic selector cache hit performance for state changes but shallowly equal selector args', () => {
@@ -382,1062 +381,5 @@ describe('Customizing selectors', () => {
     defaultMemoizeAcceptsArgFromCSC({ a: 1, b: 3 })
 
     expect(memoizer3Calls).toBeGreaterThan(0)
-  })
-})
-
-describe('defaultMemoize', () => {
-  test('Basic memoization', () => {
-    let called = 0
-    const memoized = defaultMemoize(state => {
-      called++
-      return state.a
-    })
-
-    const o1 = { a: 1 }
-    const o2 = { a: 2 }
-    expect(memoized(o1)).toBe(1)
-    expect(memoized(o1)).toBe(1)
-    expect(called).toBe(1)
-    expect(memoized(o2)).toBe(2)
-    expect(called).toBe(2)
-  })
-
-  test('Memoizes with multiple arguments', () => {
-    const memoized = defaultMemoize((...args) =>
-      args.reduce((sum, value) => sum + value, 0)
-    )
-    expect(memoized(1, 2)).toBe(3)
-    expect(memoized(1)).toBe(1)
-  })
-
-  test('Memoizes with equalityCheck override', () => {
-    // a rather absurd equals operation we can verify in tests
-    let called = 0
-    const valueEquals = (a: any, b: any) => typeof a === typeof b
-    const memoized = defaultMemoize(a => {
-      called++
-      return a
-    }, valueEquals)
-    expect(memoized(1)).toBe(1)
-    expect(memoized(2)).toBe(1) // yes, really true
-    expect(called).toBe(1)
-    expect(memoized('A')).toBe('A')
-    expect(called).toBe(2)
-  })
-
-  test('Passes correct objects to equalityCheck', () => {
-    let fallthroughs = 0
-    function shallowEqual(newVal: any, oldVal: any) {
-      if (newVal === oldVal) return true
-
-      fallthroughs += 1 // code below is expensive and should be bypassed when possible
-
-      let countA = 0
-      let countB = 0
-      for (const key in newVal) {
-        if (
-          Object.hasOwnProperty.call(newVal, key) &&
-          newVal[key] !== oldVal[key]
-        )
-          return false
-        countA++
-      }
-      for (const key in oldVal) {
-        if (Object.hasOwnProperty.call(oldVal, key)) countB++
-      }
-      return countA === countB
-    }
-
-    const someObject = { foo: 'bar' }
-    const anotherObject = { foo: 'bar' }
-    const memoized = defaultMemoize(a => a, shallowEqual)
-
-    // the first call to `memoized` doesn't hit because `defaultMemoize.lastArgs` is uninitialized
-    // and so `equalityCheck` is never called
-    memoized(someObject)
-    // first call does not shallow compare
-    expect(fallthroughs).toBe(0)
-
-    // the next call, with a different object reference, does fall through
-    memoized(anotherObject)
-
-    // call with different object does shallow compare
-    expect(fallthroughs).toBe(1)
-
-    /*
-    This test was useful when we had a cache size of 1 previously, and always saved `lastArgs`.
-    But, with the new implementation, this doesn't make sense any more.
-
-    // the third call does not fall through because `defaultMemoize` passes `anotherObject` as
-    // both the `newVal` and `oldVal` params. This allows `shallowEqual` to be much more performant
-    // than if it had passed `someObject` as `oldVal`, even though `someObject` and `anotherObject`
-    // are shallowly equal
-    memoized(anotherObject)
-    // call with same object as previous call does not shallow compare
-    expect(fallthroughs).toBe(1)
-
-    */
-  })
-
-  test('Accepts a max size greater than 1 with LRU cache behavior', () => {
-    let funcCalls = 0
-
-    const memoizer = defaultMemoize(
-      (state: any) => {
-        funcCalls++
-        return state
-      },
-      {
-        maxSize: 3
-      }
-    )
-
-    // Initial call
-    memoizer('a') // ['a']
-    expect(funcCalls).toBe(1)
-
-    // In cache - memoized
-    memoizer('a') // ['a']
-    expect(funcCalls).toBe(1)
-
-    // Added
-    memoizer('b') // ['b', 'a']
-    expect(funcCalls).toBe(2)
-
-    // Added
-    memoizer('c') // ['c', 'b', 'a']
-    expect(funcCalls).toBe(3)
-
-    // Added, removes 'a'
-    memoizer('d') // ['d', 'c', 'b']
-    expect(funcCalls).toBe(4)
-
-    // No longer in cache, re-added, removes 'b'
-    memoizer('a') // ['a', 'd', 'c']
-    expect(funcCalls).toBe(5)
-
-    // In cache, moved to front
-    memoizer('c') // ['c', 'a', 'd']
-    expect(funcCalls).toBe(5)
-
-    // Added, removes 'd'
-    memoizer('e') // ['e', 'c', 'a']
-    expect(funcCalls).toBe(6)
-
-    // No longer in cache, re-added, removes 'a'
-    memoizer('d') // ['d', 'e', 'c']
-    expect(funcCalls).toBe(7)
-  })
-
-  test('Allows reusing an existing result if they are equivalent', () => {
-    interface Todo {
-      id: number
-      name: string
-    }
-
-    const todos1: Todo[] = [
-      { id: 1, name: 'a' },
-      { id: 2, name: 'b' },
-      { id: 3, name: 'c' }
-    ]
-    const todos2 = todos1.slice()
-    todos2[2] = { id: 3, name: 'd' }
-
-    function is(x: unknown, y: unknown) {
-      if (x === y) {
-        return x !== 0 || y !== 0 || 1 / x === 1 / y
-      } else {
-        return x !== x && y !== y
-      }
-    }
-
-    function shallowEqual(objA: any, objB: any) {
-      if (is(objA, objB)) return true
-
-      if (
-        typeof objA !== 'object' ||
-        objA === null ||
-        typeof objB !== 'object' ||
-        objB === null
-      ) {
-        return false
-      }
-
-      const keysA = Object.keys(objA)
-      const keysB = Object.keys(objB)
-
-      if (keysA.length !== keysB.length) return false
-
-      for (let i = 0; i < keysA.length; i++) {
-        if (
-          !Object.prototype.hasOwnProperty.call(objB, keysA[i]) ||
-          !is(objA[keysA[i]], objB[keysA[i]])
-        ) {
-          return false
-        }
-      }
-
-      return true
-    }
-
-    for (const maxSize of [1, 3]) {
-      let funcCalls = 0
-
-      const memoizer = defaultMemoize(
-        (state: Todo[]) => {
-          funcCalls++
-          return state.map(todo => todo.id)
-        },
-        {
-          maxSize,
-          resultEqualityCheck: shallowEqual
-        }
-      )
-
-      const ids1 = memoizer(todos1)
-      expect(funcCalls).toBe(1)
-
-      const ids2 = memoizer(todos1)
-      expect(funcCalls).toBe(1)
-      expect(ids2).toBe(ids1)
-
-      const ids3 = memoizer(todos2)
-      expect(funcCalls).toBe(2)
-      expect(ids3).toBe(ids1)
-    }
-  })
-
-  test('updates the cache key even if resultEqualityCheck is a hit', () => {
-    const selector = vi.fn(x => x)
-    const equalityCheck = vi.fn((a, b) => a === b)
-    const resultEqualityCheck = vi.fn((a, b) => typeof a === typeof b)
-
-    const memoizedFn = defaultMemoize(selector, {
-      maxSize: 1,
-      resultEqualityCheck,
-      equalityCheck
-    })
-
-    // initialize the cache
-    memoizedFn('cache this result')
-    expect(selector).toBeCalledTimes(1)
-
-    // resultEqualityCheck hit (with a different cache key)
-    const result = memoizedFn('arg1')
-    expect(equalityCheck).toHaveLastReturnedWith(false)
-    expect(resultEqualityCheck).toHaveLastReturnedWith(true)
-    expect(result).toBe('cache this result')
-    expect(selector).toBeCalledTimes(2)
-
-    // cache key should now be updated
-    const result2 = memoizedFn('arg1')
-    expect(result2).toBe('cache this result')
-    expect(equalityCheck).toHaveLastReturnedWith(true)
-    expect(selector).toBeCalledTimes(2)
-  })
-
-  // Issue #527
-  test('Allows caching a value of `undefined`', () => {
-    const state = {
-      foo: { baz: 'baz' },
-      bar: 'qux'
-    }
-
-    const fooChangeSpy = vi.fn()
-
-    const fooChangeHandler = createSelector(
-      (state: any) => state.foo,
-      fooChangeSpy
-    )
-
-    fooChangeHandler(state)
-    expect(fooChangeSpy.mock.calls.length).toEqual(1)
-
-    // no change
-    fooChangeHandler(state)
-    // this would fail
-    expect(fooChangeSpy.mock.calls.length).toEqual(1)
-
-    const state2 = { a: 1 }
-    let count = 0
-
-    const selector = createSelector([(state: any) => state.a], () => {
-      count++
-      return undefined
-    })
-
-    selector(state)
-    expect(count).toBe(1)
-    selector(state)
-    expect(count).toBe(1)
-  })
-
-  test('Accepts an options object as an arg', () => {
-    let memoizer1Calls = 0
-
-    const acceptsEqualityCheckAsOption = defaultMemoize((a: any) => a, {
-      equalityCheck: (a, b) => {
-        memoizer1Calls++
-        return a === b
-      }
-    })
-
-    acceptsEqualityCheckAsOption(42)
-    acceptsEqualityCheckAsOption(43)
-
-    expect(memoizer1Calls).toBeGreaterThan(0)
-
-    let called = 0
-    const fallsBackToDefaultEqualityIfNoArgGiven = defaultMemoize(
-      state => {
-        called++
-        return state.a
-      },
-      {
-        // no args
-      }
-    )
-
-    const o1 = { a: 1 }
-    const o2 = { a: 2 }
-    expect(fallsBackToDefaultEqualityIfNoArgGiven(o1)).toBe(1)
-    expect(fallsBackToDefaultEqualityIfNoArgGiven(o1)).toBe(1)
-    expect(called).toBe(1)
-    expect(fallsBackToDefaultEqualityIfNoArgGiven(o2)).toBe(2)
-    expect(called).toBe(2)
-  })
-
-  test('Exposes a clearCache method on the memoized function', () => {
-    let funcCalls = 0
-
-    // Cache size of 1
-    const memoizer = defaultMemoize(
-      (state: any) => {
-        funcCalls++
-        return state
-      },
-      {
-        maxSize: 1
-      }
-    )
-
-    // Initial call
-    memoizer('a') // ['a']
-    expect(funcCalls).toBe(1)
-
-    // In cache - memoized
-    memoizer('a') // ['a']
-    expect(funcCalls).toBe(1)
-
-    memoizer.clearCache()
-
-    // Cache was cleared
-    memoizer('a')
-    expect(funcCalls).toBe(2)
-
-    funcCalls = 0
-
-    // Test out maxSize of 3 + exposure via createSelector
-    const selector = createSelector(
-      (state: string) => state,
-      state => {
-        funcCalls++
-        return state
-      },
-      {
-        memoizeOptions: { maxSize: 3 }
-      }
-    )
-
-    // Initial call
-    selector('a') // ['a']
-    expect(funcCalls).toBe(1)
-
-    // In cache - memoized
-    selector('a') // ['a']
-    expect(funcCalls).toBe(1)
-
-    // Added
-    selector('b') // ['b', 'a']
-    expect(funcCalls).toBe(2)
-
-    // Added
-    selector('c') // ['c', 'b', 'a']
-    expect(funcCalls).toBe(3)
-
-    // Already in cache
-    selector('c') // ['c', 'b', 'a']
-    expect(funcCalls).toBe(3)
-
-    selector.memoizedResultFunc.clearCache()
-
-    // Added
-    selector('a') // ['a']
-    expect(funcCalls).toBe(4)
-
-    // Already in cache
-    selector('a') // ['a']
-    expect(funcCalls).toBe(4)
-
-    // make sure clearCache is passed to the selector correctly
-    selector.clearCache()
-
-    // Cache was cleared
-    // Note: the outer arguments wrapper function still has 'a' in its own size-1 cache, so passing
-    // 'a' here would _not_ recalculate
-    selector('b') // ['b']
-    expect(funcCalls).toBe(5)
-
-    try {
-      //@ts-expect-error issue 591
-      selector.resultFunc.clearCache()
-      fail('should have thrown for issue 591')
-    } catch (err) {
-      //expected catch
-    }
-  })
-})
-
-describe('createStructureSelector', () => {
-  test('structured selector', () => {
-    const selector = createStructuredSelector({
-      x: (state: StateAB) => state.a,
-      y: (state: StateAB) => state.b
-    })
-    const firstResult = selector({ a: 1, b: 2 })
-    expect(firstResult).toEqual({ x: 1, y: 2 })
-    expect(selector({ a: 1, b: 2 })).toBe(firstResult)
-    const secondResult = selector({ a: 2, b: 2 })
-    expect(secondResult).toEqual({ x: 2, y: 2 })
-    expect(selector({ a: 2, b: 2 })).toBe(secondResult)
-  })
-
-  test('structured selector with invalid arguments', () => {
-    expect(() =>
-      // @ts-expect-error
-      createStructuredSelector(
-        (state: StateAB) => state.a,
-        (state: StateAB) => state.b
-      )
-    ).toThrow(/expects first argument to be an object.*function/)
-    expect(() =>
-      createStructuredSelector({
-        a: state => state.b,
-        // @ts-expect-error
-        c: 'd'
-      })
-    ).toThrow(
-      'createSelector expects all input-selectors to be functions, but received the following types: [function a(), string]'
-    )
-  })
-
-  test('structured selector with custom selector creator', () => {
-    const customSelectorCreator = createSelectorCreator(
-      defaultMemoize,
-      (a, b) => a === b
-    )
-    const selector = createStructuredSelector(
-      {
-        x: (state: StateAB) => state.a,
-        y: (state: StateAB) => state.b
-      },
-      customSelectorCreator
-    )
-    const firstResult = selector({ a: 1, b: 2 })
-    expect(firstResult).toEqual({ x: 1, y: 2 })
-    expect(selector({ a: 1, b: 2 })).toBe(firstResult)
-    expect(selector({ a: 2, b: 2 })).toEqual({ x: 2, y: 2 })
-  })
-})
-
-describe('createSelector exposed utils', () => {
-  test('resetRecomputations', () => {
-    const selector = createSelector(
-      (state: StateA) => state.a,
-      a => a
-    )
-    expect(selector({ a: 1 })).toBe(1)
-    expect(selector({ a: 1 })).toBe(1)
-    expect(selector.recomputations()).toBe(1)
-    expect(selector({ a: 2 })).toBe(2)
-    expect(selector.recomputations()).toBe(2)
-
-    selector.resetRecomputations()
-    expect(selector.recomputations()).toBe(0)
-
-    expect(selector({ a: 1 })).toBe(1)
-    expect(selector({ a: 1 })).toBe(1)
-    expect(selector.recomputations()).toBe(1)
-    expect(selector({ a: 2 })).toBe(2)
-    expect(selector.recomputations()).toBe(2)
-  })
-
-  test('export last function as resultFunc', () => {
-    const lastFunction = () => {}
-    const selector = createSelector((state: StateA) => state.a, lastFunction)
-    expect(selector.resultFunc).toBe(lastFunction)
-  })
-
-  test('export dependencies as dependencies', () => {
-    const dependency1 = (state: StateA) => {
-      state.a
-    }
-    const dependency2 = (state: StateA) => {
-      state.a
-    }
-
-    const selector = createSelector(dependency1, dependency2, () => {})
-    expect(selector.dependencies).toEqual([dependency1, dependency2])
-  })
-
-  test('export lastResult function', () => {
-    const selector = createSelector(
-      (state: StateAB) => state.a,
-      (state: StateAB) => state.b,
-      (a, b) => a + b
-    )
-
-    const result = selector({ a: 1, b: 2 })
-    expect(result).toBe(3)
-    expect(selector.lastResult()).toBe(3)
-  })
-})
-
-describe('Basic selector behavior with autotrack', () => {
-  const createSelector = createSelectorCreator(autotrackMemoize)
-
-  test('basic selector', () => {
-    // console.log('Selector test')
-    const selector = createSelector(
-      (state: StateA) => state.a,
-      a => a
-    )
-    const firstState = { a: 1 }
-    const firstStateNewPointer = { a: 1 }
-    const secondState = { a: 2 }
-
-    expect(selector(firstState)).toBe(1)
-    expect(selector(firstState)).toBe(1)
-    expect(selector.recomputations()).toBe(1)
-    expect(selector(firstStateNewPointer)).toBe(1)
-    expect(selector.recomputations()).toBe(1)
-    expect(selector(secondState)).toBe(2)
-    expect(selector.recomputations()).toBe(2)
-  })
-
-  test("don't pass extra parameters to inputSelector when only called with the state", () => {
-    const selector = createSelector(
-      (...params: any[]) => params.length,
-      a => a
-    )
-    expect(selector({})).toBe(1)
-  })
-
-  test('basic selector multiple keys', () => {
-    const selector = createSelector(
-      (state: StateAB) => state.a,
-      (state: StateAB) => state.b,
-      (a, b) => a + b
-    )
-    const state1 = { a: 1, b: 2 }
-    expect(selector(state1)).toBe(3)
-    expect(selector(state1)).toBe(3)
-    expect(selector.recomputations()).toBe(1)
-    const state2 = { a: 3, b: 2 }
-    expect(selector(state2)).toBe(5)
-    expect(selector(state2)).toBe(5)
-    expect(selector.recomputations()).toBe(2)
-  })
-
-  test('basic selector invalid input selector', () => {
-    expect(() =>
-      createSelector(
-        // @ts-ignore
-        (state: StateAB) => state.a,
-        function input2(state: StateAB) {
-          return state.b
-        },
-        'not a function',
-        (a: any, b: any) => a + b
-      )
-    ).toThrow(
-      'createSelector expects all input-selectors to be functions, but received the following types: [function unnamed(), function input2(), string]'
-    )
-
-    expect(() =>
-      // @ts-ignore
-      createSelector((state: StateAB) => state.a, 'not a function')
-    ).toThrow(
-      'createSelector expects an output function after the inputs, but received: [string]'
-    )
-  })
-
-  test('basic selector cache hit performance', () => {
-    if (process.env.COVERAGE) {
-      return // don't run performance tests for coverage
-    }
-
-    const selector = createSelector(
-      (state: StateAB) => state.a,
-      (state: StateAB) => state.b,
-      (a, b) => a + b
-    )
-    const state1 = { a: 1, b: 2 }
-
-    const start = performance.now()
-    for (let i = 0; i < 1000000; i++) {
-      selector(state1)
-    }
-    const totalTime = performance.now() - start
-    console.log('Total time', totalTime)
-
-    expect(selector(state1)).toBe(3)
-    expect(selector.recomputations()).toBe(1)
-    // Expected a million calls to a selector with the same arguments to take less than 1 second
-    expect(totalTime).toBeLessThan(1000)
-  })
-
-  test('basic selector cache hit performance for state changes but shallowly equal selector args', () => {
-    if (process.env.COVERAGE) {
-      return // don't run performance tests for coverage
-    }
-
-    const selector = createSelector(
-      (state: StateAB) => state.a,
-      (state: StateAB) => state.b,
-      (a, b) => a + b
-    )
-
-    const start = performance.now()
-    for (let i = 0; i < 1000000; i++) {
-      selector(states[i])
-    }
-    const totalTime = performance.now() - start
-
-    console.log('Total time', totalTime)
-
-    expect(selector(states[0])).toBe(3)
-    expect(selector.recomputations()).toBe(1)
-
-    // Expected a million calls to a selector with the same arguments to take less than 1 second
-    expect(totalTime).toBeLessThan(1000)
-  })
-
-  test('memoized composite arguments', () => {
-    const selector = createSelector(
-      (state: StateSub) => state.sub,
-      sub => sub.a
-    )
-    const state1 = { sub: { a: 1 } }
-    expect(selector(state1)).toEqual(1)
-    expect(selector(state1)).toEqual(1)
-    expect(selector.recomputations()).toBe(1)
-    const state2 = { sub: { a: 2 } }
-    expect(selector(state2)).toEqual(2)
-    expect(selector.recomputations()).toBe(2)
-  })
-
-  test('first argument can be an array', () => {
-    const selector = createSelector(
-      [state => state.a, state => state.b],
-      (a, b) => {
-        return a + b
-      }
-    )
-    expect(selector({ a: 1, b: 2 })).toBe(3)
-    expect(selector({ a: 1, b: 2 })).toBe(3)
-    expect(selector.recomputations()).toBe(1)
-    expect(selector({ a: 3, b: 2 })).toBe(5)
-    expect(selector.recomputations()).toBe(2)
-  })
-
-  test('can accept props', () => {
-    let called = 0
-    const selector = createSelector(
-      (state: StateAB) => state.a,
-      (state: StateAB) => state.b,
-      (state: StateAB, props: { c: number }) => props.c,
-      (a, b, c) => {
-        called++
-        return a + b + c
-      }
-    )
-    expect(selector({ a: 1, b: 2 }, { c: 100 })).toBe(103)
-  })
-
-  test('recomputes result after exception', () => {
-    let called = 0
-    const selector = createSelector(
-      (state: StateA) => state.a,
-      () => {
-        called++
-        throw Error('test error')
-      }
-    )
-    expect(() => selector({ a: 1 })).toThrow('test error')
-    expect(() => selector({ a: 1 })).toThrow('test error')
-    expect(called).toBe(2)
-  })
-
-  test('memoizes previous result before exception', () => {
-    let called = 0
-    const selector = createSelector(
-      (state: StateA) => state.a,
-      a => {
-        called++
-        if (a > 1) throw Error('test error')
-        return a
-      }
-    )
-    const state1 = { a: 1 }
-    const state2 = { a: 2 }
-    expect(selector(state1)).toBe(1)
-    expect(() => selector(state2)).toThrow('test error')
-    expect(selector(state1)).toBe(1)
-    expect(called).toBe(2)
-  })
-})
-
-describe('More perf comparisons', () => {
-  const csDefault = createSelectorCreator(defaultMemoize)
-  const csAutotrack = createSelectorCreator(autotrackMemoize)
-
-  interface Todo {
-    id: number
-    name: string
-    completed: boolean
-  }
-
-  type TodosState = Todo[]
-
-  const counterSlice = createSlice({
-    name: 'counters',
-    initialState: {
-      deeply: {
-        nested: {
-          really: {
-            deeply: {
-              nested: {
-                c1: { value: 0 }
-              }
-            }
-          }
-        }
-      },
-
-      c2: { value: 0 }
-    },
-    reducers: {
-      increment1(state) {
-        // state.c1.value++
-        state.deeply.nested.really.deeply.nested.c1.value++
-      },
-      increment2(state) {
-        state.c2.value++
-      }
-    }
-  })
-
-  const todosSlice = createSlice({
-    name: 'todos',
-    initialState: [
-      { id: 0, name: 'a', completed: false },
-      { id: 1, name: 'b', completed: false },
-      { id: 2, name: 'c', completed: false }
-    ] as TodosState,
-    reducers: {
-      toggleCompleted(state, action: PayloadAction<number>) {
-        const todo = state.find(todo => todo.id === action.payload)
-        if (todo) {
-          todo.completed = !todo.completed
-        }
-      },
-      setName(state) {
-        state[1].name = 'd'
-      }
-    }
-  })
-
-  const store = configureStore({
-    reducer: {
-      counter: counterSlice.reducer,
-      todos: todosSlice.reducer
-    },
-    middleware: gDM =>
-      gDM({
-        serializableCheck: false,
-        immutableCheck: false
-      })
-  })
-
-  type RootState = ReturnType<typeof store.getState>
-
-  const states: RootState[] = []
-
-  for (let i = 0; i < 1000; i++) {
-    states.push(store.getState())
-    store.dispatch(counterSlice.actions.increment1())
-    states.push(store.getState())
-    store.dispatch(counterSlice.actions.increment2())
-    states.push(store.getState())
-    store.dispatch(todosSlice.actions.toggleCompleted(1))
-    states.push(store.getState())
-    store.dispatch(todosSlice.actions.setName())
-    states.push(store.getState())
-  }
-
-  it('More detailed perf comparison', () => {
-    const cdCounters1 = csDefault(
-      (state: RootState) =>
-        state.counter.deeply.nested.really.deeply.nested.c1.value,
-      (state: RootState) => state.counter.c2.value,
-      (c1, c2) => {
-        return c1 + c2
-      }
-    )
-
-    const cdCounters2 = csDefault(
-      (state: RootState) => state.counter.deeply.nested.really.deeply.nested.c1,
-      (state: RootState) => state.counter.c2,
-      (c1, c2) => {
-        return c1.value + c2.value
-      }
-    )
-
-    const cdTodoIds = csDefault(
-      (state: RootState) => state.todos,
-      todos => {
-        return todos.map(todo => todo.id)
-      }
-    )
-
-    const cdTodoIdsAndNames = csDefault(
-      (state: RootState) => state.todos,
-      todos => {
-        return todos.map(todo => ({ id: todo.id, name: todo.name }))
-      }
-    )
-
-    const cdCompletedTodos = csDefault(
-      (state: RootState) => state.todos,
-      todos => {
-        const completed = todos.filter(todo => todo.completed)
-        return completed.length
-      }
-    )
-
-    const cdCompletedTodos2 = csDefault(
-      (state: RootState) => state.todos,
-      todos => {
-        const completed = todos.filter(todo => todo.completed)
-        return completed.length
-      }
-    )
-
-    const caCounters1 = csDefault(
-      (state: RootState) =>
-        state.counter.deeply.nested.really.deeply.nested.c1.value,
-      (state: RootState) => state.counter.c2.value,
-      (c1, c2) => {
-        return c1 + c2
-      }
-    )
-
-    const caCounters2 = csAutotrack(
-      (state: RootState) => state.counter.deeply.nested.really.deeply.nested.c1,
-      (state: RootState) => state.counter.c2,
-      (c1, c2) => {
-        // console.log('inside caCounters2: ', { c1, c2 })
-        return c1.value + c2.value
-      }
-    )
-
-    const caTodoIds = csAutotrack(
-      (state: RootState) => state.todos,
-      todos => {
-        return todos.map(todo => todo.id)
-      }
-    )
-
-    const caTodoIdsAndNames = csAutotrack(
-      (state: RootState) => state.todos,
-      todos => {
-        return todos.map(todo => ({ id: todo.id, name: todo.name }))
-      }
-    )
-
-    const caCompletedTodos = csAutotrack(
-      (state: RootState) => state.todos,
-      todos => {
-        const completed = todos.filter(todo => todo.completed)
-        return completed.length
-      }
-    )
-
-    const caCompletedTodos2 = csAutotrack(
-      (state: RootState) => state.todos,
-      todos => {
-        const completed = todos.filter(todo => todo.completed)
-        return completed.length
-      }
-    )
-
-    const defaultStart = performance.now()
-    for (const state of states) {
-      cdCounters1(state)
-      cdCounters2(state)
-      // console.log('csCounters2', cdCounters2(state))
-      cdTodoIds(state)
-      cdTodoIdsAndNames(state)
-      cdCompletedTodos(state)
-      cdCompletedTodos2(state)
-    }
-    const defaultEnd = performance.now()
-
-    const autotrackStart = performance.now()
-    for (const state of states) {
-      caCounters1(state)
-      caCounters2(state)
-      // console.log('State.counter: ', state.counter)
-      // console.log('caCounters2', caCounters2(state))
-      caTodoIds(state)
-      caTodoIdsAndNames(state)
-      caCompletedTodos(state)
-      caCompletedTodos2(state)
-    }
-    const autotrackEnd = performance.now()
-
-    const allSelectors = {
-      cdCounters1,
-      cdCounters2,
-      cdTodoIds,
-      cdTodoIdsAndNames,
-      cdCompletedTodos,
-      cdCompletedTodos2,
-      caCounters1,
-      caCounters2,
-      caTodoIds,
-      caTodoIdsAndNames,
-      caCompletedTodos,
-      caCompletedTodos2
-    }
-
-    console.log('\nTotal recomputations:')
-    Object.entries(allSelectors).forEach(([name, selector]) => {
-      console.log(name, selector.recomputations())
-    })
-
-    console.log('Total elapsed times: ', {
-      defaultElapsed: defaultEnd - defaultStart,
-      autotrackElapsed: autotrackEnd - autotrackStart
-    })
-  })
-
-  it.skip('weakMapMemoizer recalcs', () => {
-    const state1 = store.getState()
-
-    store.dispatch(counterSlice.actions.increment1())
-    const state2 = store.getState()
-
-    const csWeakmap = createSelectorCreator(weakMapMemoize)
-
-    const cwCounters2 = csWeakmap(
-      (state: RootState) => state.counter.deeply.nested.really.deeply.nested.c1,
-      (state: RootState) => state.counter.c2,
-      (c1, c2) => {
-        // console.log('inside caCounters2: ', { c1, c2 })
-        return c1.value + c2.value
-      }
-    )
-
-    for (let i = 0; i < 10; i++) {
-      cwCounters2(state1)
-      cwCounters2(state2)
-    }
-
-    console.log('cwCounters2.recomputations()', cwCounters2.recomputations())
-  })
-
-  test.only('Does something?', async () => {
-    const fn = vi.fn()
-
-    let resolve: () => void
-    const promise = new Promise<void>(r => (resolve = r))
-
-    const registry = new FinalizationRegistry(heldValue => {
-      resolve()
-      console.log('Garbage-collected value for ID: ', heldValue)
-      fn(heldValue)
-    })
-
-    const createSelectorWeakmap = createSelectorCreator(weakMapMemoize)
-
-    const store = configureStore({
-      reducer: {
-        counter: counterSlice.reducer,
-        todos: todosSlice.reducer
-      },
-      middleware: gDM =>
-        gDM({
-          serializableCheck: false,
-          immutableCheck: false
-        })
-    })
-
-    const reduxStates: RootState[] = []
-
-    const NUM_ITEMS = 10
-
-    for (let i = 0; i < NUM_ITEMS; i++) {
-      store.dispatch(todosSlice.actions.toggleCompleted(1))
-      const state = store.getState()
-      reduxStates.push(state)
-      registry.register(state, i)
-    }
-
-    const cdTodoIdsAndNames = createSelectorWeakmap(
-      (state: RootState) => state.todos,
-      todos => {
-        // console.log('Recalculating todo IDs')
-        return todos.map(todo => ({ id: todo.id, name: todo.name }))
-      }
-    )
-
-    for (const state of reduxStates) {
-      cdTodoIdsAndNames(state)
-    }
-
-    expect(cdTodoIdsAndNames.recomputations()).toBe(NUM_ITEMS)
-
-    for (const state of reduxStates) {
-      cdTodoIdsAndNames(state)
-    }
-
-    expect(cdTodoIdsAndNames.recomputations()).toBe(NUM_ITEMS)
-
-    console.log('clearCache: ', cdTodoIdsAndNames.clearCache)
-    cdTodoIdsAndNames.memoizedResultFunc.clearCache()
-
-    cdTodoIdsAndNames(reduxStates[0])
-
-    expect(cdTodoIdsAndNames.recomputations()).toBe(NUM_ITEMS + 1)
-
-    cdTodoIdsAndNames(reduxStates[1])
-
-    expect(cdTodoIdsAndNames.recomputations()).toBe(NUM_ITEMS + 2)
-
-    console.log('Before nulling out')
-    // @ts-ignore
-    reduxStates[0] = null
-    console.log('After nulling out')
-    if (global.gc) {
-      global.gc()
-    }
-    console.log('After GC')
-
-    await promise
-    expect(fn).toHaveBeenCalledWith(0)
-
-    // garbage-collected for ID: 3
   })
 })
