+++ test/testTypes.ts
@@ -1,0 +1,14 @@
+export interface StateA {
+  a: number
+}
+
+export interface StateAB {
+  a: number
+  b: number
+}
+
+export interface StateSub {
+  sub: {
+    a: number
+  }
+}
